railway_spray_widget/requirements.txt
flask
gunicorn
pandas
numpy
matplotlib
requests
python-dotenv
